requirements.txt
# FastAPI and Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
websockets==13.1

# LangChain Framework
langchain==0.1.4
langchain-community==0.0.16
langchain-openai==0.0.5
langchain-anthropic==0.1.1
langsmith==0.0.83

# Database / Supabase Stack
supabase==2.21.1
postgrest==2.21.1
gotrue==2.12.4
storage3==2.21.1
realtime==2.21.1
supabase-auth==2.21.1
supabase-functions==2.21.1

# Environment & Configuration
python-dotenv==1.0.0
pydantic==2.11.10
pydantic-settings==2.1.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
email-validator==2.1.0

# External API Clients
httpx==0.28.1
aiohttp==3.9.1

# YAML Processing
pyyaml==6.0.1

# Text normalization
unidecode==1.4.0

# Airport/City code mapping (custom implementation)

# Date/Time
python-dateutil==2.8.2

# Background Jobs
celery==5.3.4
redis==5.0.1

# Session Management
fastapi-sessions==0.3.2

# Utilities
tiktoken==0.5.2  # Token counting for context management
tenacity==8.2.3  # Retry logic
unidecode==1.4.0

# Development & Testing
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.1
flake8==7.0.0
mypy==1.8.0

# Monitoring (Optional)
prometheus-client==0.19.0
